+++ tests/database/factories/ACLFactory.php
@@ -4,7 +4,7 @@ use Tests\Utils\Models\ACL;
 use Faker\Generator as Faker;
 
 /* @var \Illuminate\Database\Eloquent\Factory $factory */
-$factory->define(ACL::class, function (Faker $faker) {
+$factory->define(ACL::class, function (Faker $faker): array {
     return [
         'create_post' => $faker->boolean,
         'read_post' => $faker->boolean,